Replaced type "int" with int8_t/int16_t/int32_t (coding conventions). Renamed RPi3 build target to align with architecture.

+++ command.c
@@ -1269,7 +1269,7 @@ void uart_calc_crc(command_t * cmd) {
 	cmd->to = crc & 0xff;
 #if defined PC && 0
 	/* CRC Test */
-	int r = rand();
+	int8_t r = rand();
 	if (r > RAND_MAX * 0.999) {
 		cmd->to++;
 	}
